working booking page

+++ MvcKissApplication/Api/Controllers/AdministratorController.cs
@@ -3,10 +3,12 @@ using MvcKissApplication.Api.ViewModels;
 using MvcKissApplication.Database.Models;
 using System;
 using System.Collections.Generic;
+using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Web.Http;
+using WebAPI.OutputCache;
 
 namespace MvcKissApplication.Api.controllers
 {
@@ -30,8 +32,10 @@ namespace MvcKissApplication.Api.controllers
 
         #endregion CONSTRUCTORS
 
+        // GET api/administrator
         [HttpGet]
         [ActionName("DefaultAction")]
+        [CacheOutput(ClientTimeSpan = 0, ServerTimeSpan = 0)]
         public IEnumerable<AdministratorView> Get()
         {
             var models = repo.getAllAdministrators();
@@ -39,8 +43,10 @@ namespace MvcKissApplication.Api.controllers
             return views;
         }
 
+        // GET api/administrator/5
         [HttpGet]
         [ActionName("DefaultAction")]
+        [CacheOutput(ClientTimeSpan = 0, ServerTimeSpan = 0)]
         public HttpResponseMessage Get(int id)
         {
             var model = repo.getAdministrator(id);
@@ -70,6 +76,7 @@ namespace MvcKissApplication.Api.controllers
             }
         }
 
+        // POST api/administrator
         [HttpPost]
         [ActionName("DefaultAction")]
         public HttpResponseMessage Post(AdministratorView view)
@@ -77,31 +84,69 @@ namespace MvcKissApplication.Api.controllers
             var model = view.getModel();
             model.created = DateTime.UtcNow;
             model.updated = DateTime.UtcNow;
+            model.fakeId = Guid.NewGuid();
             model = repo.createAdministrator(model);
             view = new AdministratorView(model);
 
-            var response = Request.CreateResponse<AdministratorView>(HttpStatusCode.Created, view);
-            string uri = Url.Route(null, new { id = view.id });
-            response.Headers.Location = new Uri(Request.RequestUri, uri);
+            HttpResponseMessage response = Request.CreateResponse(HttpStatusCode.Created, view);
+            response.Headers.Location = new Uri(Url.Link("ApiControllerAndId", new { id = view.id }));
             return response;
         }
 
+        // PUT api/administrator/5
         [HttpPut]
         [ActionName("DefaultAction")]
-        public void Put(int id, AdministratorView view)
+        public HttpResponseMessage Put(int id, AdministratorView view)
         {
-            view.id = id;
+            if (id != view.id)
+            {
+                return Request.CreateResponse(HttpStatusCode.BadRequest);
+            }
+            
             var model = view.getModel();
             model.updated = DateTime.UtcNow;
-            repo.update(model);
+
+            try
+            {
+                repo.update(model);
+            }
+            catch (DbUpdateConcurrencyException ex)
+            {
+                return Request.CreateErrorResponse(HttpStatusCode.NotFound, ex);
+            }
+
+            return Request.CreateResponse(HttpStatusCode.OK);
         }
 
+        // DELETE api/admin/5
         [HttpDelete]
         [ActionName("DefaultAction")]
-        public void Remove(int id)
+        public HttpResponseMessage Remove(int id)
         {
-            repo.deleteAdministrator(id);
+            var model = repo.getAdministrator(id);
+            if (model == null)
+            {
+                return Request.CreateResponse(HttpStatusCode.NotFound);
+            }
+            else
+            {
+                try 
+                {
+                    repo.deleteAdministrator(id);
+                }
+                catch (DbUpdateConcurrencyException ex)
+                {
+                    return Request.CreateErrorResponse(HttpStatusCode.NotFound, ex);
+                }
+
+                return Request.CreateResponse(HttpStatusCode.OK, model);
+            }
         }
 
+        protected override void Dispose(bool disposing)
+        {
+            repo.dispose();
+            base.Dispose(disposing);
+        }
     }
 }
